backend/requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
celery==5.5.3
redis==7.0.1
elasticsearch==8.16.0
pandas==2.3.3
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
requests==2.32.5
reportlab==4.4.4
matplotlib==3.10.7
weasyprint==66.0
pydantic==2.12.3
pydantic-settings==2.11.0
pyarrow==17.0.0
httpx==0.27.0
email-validator==2.1.0
duckdb==1.1.3
psycopg2-binary==2.9.9
